### 2020-07-25 #2519 @whedon
[PRE REVIEW]: The Dusty Evolved Star Kit (DESK): A Python package for fitting the Spectral Energy Distribution of Evolved Stars
**Submitting author:** @s-goldman (<a href="http://orcid.org/0000-0002-8937-3844">Steve Goldman</a>)
**Repository:** <a href="https://github.com/s-goldman/Dusty-Evolved-Star-Kit" target ="_blank">https://github.com/s-goldman/Dusty-Evolved-Star-Kit</a>
**Version:** v1.7.0
