### 2020-07-25 #2519 @Kevin-Mattheus-Moerman
@harpolea @xuanxu @mbobra @pibion Could one of you handle this submission? Thanks. This package appears to offer novel functionality but it does seem on the minor side. If you pick this up can you also judge if this is in scope (use `@whedon query scope` to trigger an editorial board level review if needed).
