### 2020-07-25 #2519 @arfon
@s-goldman - does this submission accompany a paper in a AAS journal?
