### 2020-07-25 #2519 @s-goldman
@arfon It does not. I will, however, discuss the package and do in-depth comparisons of models for the project/paper I will be working on for the bulk of next year.
