### 2020-07-26 #2519 @arfon
@s-goldman - this submission is on the smaller-side of what we are allowing in JOSS these days and I note the following comment in your submission message to JOSS:

> The next iteration will develop a Bayesian-fitting approach, currently under development in the probabilities module that is not used in v1.7.0.

Could you say a little more about the timeline for developing this? I could imagine that the addition of this functionality would make this a more compelling submission here.
