### 2020-07-27 #2519 @s-goldman
@arfon Yeah sure, I developed that Bayesian-fitting functionality, but I got stuck on what to give the user as the default priors given that they're not well constrained. I was hoping to constrain them over the next year (using data from PHAT) and present them with the Bayesian-fitting module with my next paper.

I think that the current brute-force least-squares fitting method is the more adaptable for the average user, with a lower possibility of unrealistic results. That's at least the reason why I decided to leave it out for the time being. I suppose I could enable the package assuming flat priors, but that would have the same effect as the least-squares fit.

My timeline "plan" was have a well-established package before JWST launch, when data (other than archival Spitzer data) will finally be available to use with the package.
