### 2020-08-04 #2519 @arfon
@s-goldman - thanks for providing the additional context. Could you suggest a few people that might be able to give this package a good review? Please try to avoid tagging them by username, e.g. `@arfon` rather than @arfon.

This spreadsheet of volunteers might be useful too: https://bit.ly/joss-reviewers
