### 2020-08-04 #2519 @s-goldman
@arfon Given their background in stellar physics I would suggest `@TomGoffrey, @Deech08, or @zpace`.
