### 2020-08-04 #2519 @arfon
@schristophe @TomGoffrey @Deech08 - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
