### 2020-08-04 #2519 @Deech08
This looks really interesting, but I am defending my thesis next week and don't have the time right away to put into this. If a timeline of me starting to review in 1.5-2 weeks from now is okay, then I can. Otherwise it may be best to look for someone else.
