### 2020-08-04 #2519 @arfon
> This looks really interesting, but I am defending my thesis next week and don't have the time right away to put into this. If a timeline of me starting to review in 1.5-2 weeks from now is okay, then I can. Otherwise it may be best to look for someone else.

Absolutely! We generally ask for reviews to be complete within six weeks. If that sounds OK, I'll go ahead and add you as a reviewer (there will be two or more reviewers) and then check in with you a little later on.

>  I am defending my thesis next week

Also, good luck!
