### 2020-08-05 a874948
minor bug fixes to data retrieval script
