+++ desk/probabilities/compute_grid_weights.py
@@ -27,7 +27,7 @@ def compute_bin_boundaries(tab):
     Note
     ----
     The bin boundaries are defined as the midpoint between each value in tab.
-    At the two edges, 1/2 of the bin width is subtractted/added to the
+    At the two edges, 1/2 of the bin width is subtracted/added to the
     min/max of tab.
     """
     temp = tab[1:] - np.diff(tab) / 2.0
