### 2020-08-05 bebd842
added documentationfor compute_grid_weights.py
