### 2020-08-05 a98f2d6
updates plotting scripts
