+++ desk/outputs/plotting_seds.py
@@ -1,12 +1,82 @@
 import os
+import sys
 import ipdb
 import math
+import seaborn as sns
 import astropy.units as u
 import matplotlib.pyplot as plt
 import numpy as np
 from fnmatch import fnmatch
 from astropy.table import Table
 from desk.set_up import get_data
+
+sns.set_palette("colorblind")
+
+
+def plot_phot(x_data, y_data, ax):
+    ax.scatter(x_data, y_data)
+    return ax
+
+
+def plot_model(x_model, y_model, ax):
+    ax.plot(x_model, y_model, c="k", linewidth=0.4, linestyle="--", zorder=2)
+    return ax
+
+
+def source_name_annotation(name, ax):
+    ax.annotate(
+        str(name).replace("_", " "),
+        (0.95, 0.8),
+        xycoords="axes fraction",
+        ha="right",
+        fontsize=14,
+    )
+
+
+def counter_annotations(counter, ax):
+    ax.annotate(
+        str(counter + 1),
+        (0.075, 0.85),
+        ha="left",
+        va="center",
+        xycoords="axes fraction",
+        fontsize=14,
+    )
+    return ax
+
+
+def set_inward_ticks(ax):
+    ax.get_xaxis().set_tick_params(which="both", direction="in", labelsize=12)
+    ax.get_yaxis().set_tick_params(which="both", direction="in", labelsize=12)
+    return ax
+
+
+def set_limits(x_model, y_model, x_data, y_data, ax):
+    # set y_limits using median of model in wavelength range of data
+    median = np.median(y_model[(x_model > np.min(x_data)) & (x_model < np.max(x_data))])
+    y_min = median - 2  # room below
+    y_max = median + 1.75  # room above
+    # addes to range if median data is out of range
+    y_diff = np.median(y_data) - median
+    if y_diff > 0:
+        y_max = y_max + y_diff
+    else:
+        y_min = y_min + y_diff
+    ax.set_xlim(-0.99, 2.49)
+    ax.set_ylim(y_min, y_max)
+    return ax
+
+
+def add_axis_labels(fig, fontsize):
+    # Common figure labels
+    ax = fig.add_subplot(111, frameon=False)
+    ax.tick_params(labelcolor="none", top=False, bottom=False, left=False, right=False)
+    ax.grid(False)
+    ax.set_xlabel(r"log $\lambda$ ($\mu m$)", labelpad=10, fontsize=fontsize)
+    ax.set_ylabel(
+        r"log $\lambda$ F$_{\lambda}$ " + "(W m$^{-2}$)", labelpad=15, fontsize=fontsize
+    )
+    return fig
 
 
 def get_model_and_data_for_plotting(counter, target, source_path, source_filename):
@@ -64,6 +134,46 @@ def get_model_and_data_for_plotting(counter, target, source_path, source_filenam
     return x_model, y_model, x_data, y_data
 
 
+def single_figures(source_path, source_filename, dest_path):
+    """
+    Takes results from fitting_plotting_outputs.csv and plots SEDs.
+    Plots in individual figures.
+
+    Returns
+    -------
+    png's
+        SED figures with data in blue and model in black.
+
+    """
+    input_file = Table.read(source_path + "/" + source_filename)
+
+    for counter, target in enumerate(input_file):
+        # gets data for plotting
+        x_model, y_model, x_data, y_data = get_model_and_data_for_plotting(
+            counter, target, source_path, source_filename
+        )
+
+        # Figure plotting
+        fig, ax1 = plt.subplots(1, 1, sharex=True, sharey=True, figsize=(6, 4))
+        set_limits(x_model, y_model, x_data, y_data, ax1)
[…]
