### 2020-08-05 ce6f3fb
added seaborn as requirement
