### 2020-08-06 #2519 @TomGoffrey
@arfon Sure happy to help out. I'm ~ a day from finishing another review then I can start on this.

I do have a couple of thoughts on the discussion above, but perhaps it's best I wait until the reviewing process formally starts?
