### 2020-08-06 #2554 @TomGoffrey
Some comments from my first quick run through you might want to consider:

1. Consider replacing 'BSD License' with '3-clause BSD License' or 'BSD-3-Clause' or something a little more specific in the `LICENSE` file.
2. Is there a complete example contained within the repository anywhere? I see lots of comments similar to `--source='target_name.csv'` in the docs. Could such a csv file be include?
3. In `CONTRIBUTING.rst` there's a suggestion to use flake8 to check the code, but a cursory check of a few files suggests the code isn't fully PEP8 compliant. I've not used flake8 so perhaps I have miss understood something.
