### 2020-08-06 #2554 @s-goldman
@TomGoffrey Thanks for the comments, and so quickly.

1. Can do
2. I can be more explicit in the documentation, but running the code `desk fit` runs the code using the 3 examples sources in the `put_target_data_here` directory.
3. That should be updated to say just to use pytest. Style checks are done by Codacy after a pull request. Thank you for catching that.

All really helpful comments, and I look forward to more. Thanks Tom.
