### 2020-08-06 e39809a
Added Tom's 3 suggestions for the licence, explicit instruction on how to use the examples, and contributing guidlines.
