### 2020-08-11 ca0dc53
More robust method of saving the fit results of any model grid.
