### 2020-08-11 32fdb44
More robust method of saving the fit results of any model grid.
