+++ desk/set_up/create_output_files.py
@@ -1,6 +1,5 @@
 # Steve Goldman, Space Telescope Science Institute, sgoldman@stsci.edu
 import os
-import ipdb
 from desk.set_up import config
 
 
