### 2020-08-12 1fbb472
Description of new desk-mix grid in the grids section of the documentation.
