### 2020-08-12 2584aac
Description of new desk-mix grid in the grids section of the documentation.
