### 2020-08-12 #2554 @TomGoffrey
Looking at the paper I have a couple of general comments:

1. I think you should list all the available grids, with the appropriate citations, as you've done on in the online documentation. A table format might be appropriate (name, size, description + citation)?
2. The included figures would be a lot more useful if you included a description of what they are and how they were generated using desk. I think the latter is already included in the online documentation also.

Sorry this review is a little drip fed, a necessity of time constraints unfortunately.
