### 2020-08-12 #2554 @Deech08
> @Deech08, @TomGoffrey - please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> Any questions/concerns please let me know.

@arfon I'm having trouble getting "edit" access to the checklist. The link to accept the invitation seemed to have expired on my end (sorry I didn't get around to doing that sooner).
