### 2020-08-12 #2554 @arfon
> @arfon I'm having trouble getting "edit" access to the checklist. The link to accept the invitation seemed to have expired on my end (sorry I didn't get around to doing that sooner).

No problem. These unfortunately expire these days but you should be able to accept the invite again now ☝️
