### 2020-08-12 #2554 @TomGoffrey
@s-goldman Could I also check if desk has been used in any previous publications? If so, these should be listed as per the guidance on paper contents, and would obviously strengthen the statement of need.

If it hasn't been used, that's certainly not a problem, but some discussion of how this work fits in with the wider context would be helpful. For example, are there large data sets that could be handled by desk?
