### 2020-08-13 718d3ec
Description of new desk-mix grid in the grids section of the documentation.
