### 2020-08-13 #2554 @s-goldman
> @s-goldman Could I also check if desk has been used in any previous publications? If so, these should be listed as per the guidance on paper contents, and would obviously strengthen the statement of need.
>

The method has been used in each of my previous first author publications and [Orosz et al. 2017](https://ui.adsabs.harvard.edu/abs/2017AJ....153..119O/abstract), but not mentioned explicitly by name. It was also mentioned in [Nanni et al. 2019](https://ui.adsabs.harvard.edu/abs/2019MNRAS.487..502N/abstract), but results from the code were not used. I'm not sure if these count?

> If it hasn't been used, that's certainly not a problem, but some discussion of how this work fits in with the wider context would be helpful. For example, are there large data sets that could be handled by desk?

I have used the package on sets of tens of thousands of sources detected with the PHAT survey. In theory you could fit as many sources as you want. This will be huge with the launch of the James Webb Space Telescope, when we have a flood of new IR data. As these stars are also variable, the LSST will also identify large samples of evolved stars.
