### 2020-08-13 9337d57
Description of new desk-mix grid in the grids section of the documentation.
