### 2020-08-13 0e74b86
Updated JOSS paper text with beter figure description.
