### 2020-08-13 76d5d0e
Updated JOSS paper text with beter figure description.
