### 2020-08-13 c07e8fe
Updated JOSS paper text with beter figure description.
