### 2020-08-13 299c3b0
Updated JOSS paper text with beter figure description.
