+++ paper.md
@@ -33,17 +33,17 @@ Table 1.
 | Grid name           	| Size   	| Type 	| Atmospheric model 	| Optical constants                                          	| References                                               	|
 |---------------------	|--------	|------	|-------------------	|------------------------------------------------------------	|----------------------------------------------------------	|
 | Oss-Orich-aringer   	| 2,000  	| O    	| COMARCS           	| Warm silicates                                             	| [@Ossenkopf:1992, @Aringer:2016]                        	|
-| Oss-Orich-bb        	| 2,000  	| O    	| Black body (BB)   	| Warm silicates                                             	| 1,3                                                      	|
+| Oss-Orich-bb        	| 2,000  	| O    	| Black body (BB)   	| Warm silicates                                             	| [@Ossenkopf:1992]                                     	|
 | Crystalline-20-bb   	| 2,000  	| O    	| BB                	| 80% warm silicates 20% crystalline silicates               	| 1,3                                                      	|
-| corundum-20-bb      	| 2,000  	| O    	| BB                	| 80% warm silicates 20% corundum silicates                  	|                                                          	|
-| big-grain           	| 2,000  	| O    	| BB                	| Warm silicates with higher maximum dust grain size of 0.35 	|                                                          	|
+| corundum-20-bb      	| 2,000  	| O    	| BB                	| 80% warm silicates 20% corundum grains                    	|                                                          	|
+| big-grain           	| 2,000  	| O    	| BB                	| Warm silicates with higher maximum dust grain size           	|                                                          	|
 | fifth-iron          	| 500    	| O    	| BB                	| 80% warm silicates 20% iron grains                         	|                                                          	|
 | half-iron           	| 500    	| O    	| BB                	| 50% warm silicates 50% iron grains                         	|                                                          	|
 | one-fifth-carbon    	| 500    	| O    	| BB                	| 80% warm silicates 20% carbonaceous grains                 	|                                                          	|
 | arnold-palmer       	| 500    	| O    	| BB                	| 50% warm silicates 50% carbonaceous grains                 	|                                                          	|
 | desk-mix            	| xx,xxx 	| O    	| MARCS             	| Mixture of olivine(48-88%), corundum(8-48%), and iron (4%) 	|                                                          	|
-| Zubko-Crich-aringer 	| 2,000  	| C    	| COMARCS           	| amorphous carbon grains                                    	|                                                          	|
-| Zubko-Crich-bb      	| 2,000  	| C    	| BB                	| amorphous carbon grains                                    	|                                                          	|
+| Zubko-Crich-aringer 	| 2,000  	| C    	| COMARCS           	| Amorphous carbon grains                                    	|                                                          	|
+| Zubko-Crich-bb      	| 2,000  	| C    	| BB                	| Amorphous carbon grains                                    	|                                                          	|
 | H11-LMC             	| 90,899 	| C    	| COMARCS           	| Dust-growth grid with 1/2 solar metallicity                	|                                                          	|
 | H11-SMC             	| 91,058 	| C    	| COMARCS           	| Dust-growth grid with 1/5 solar metallicity                	|                                                          	|
 | J1000-LMC           	| 85,392 	| C    	| COMARCS           	| Dust-growth grid with 1/2 solar metallicity                	|                                                          	|
