+++ paper.md
@@ -24,12 +24,22 @@ One of the few ways that we can understand the environment around dusty stars an
 
 The ``DESK`` is a python package that fits photometry or spectra of evolved stars, or the Spectral Energy Distribution (SED), to grids of radiative transfer models using a least-squares method. The package includes newly created grids using a variety of different dust species, and state-of-the-art dust growth grids [@Nanni:2019].
 
-Results from these grids (e.g. luminosity, mass-loss rate) can vary dramatically as a result of the unknown properties of evolved stars, especially the oxygen-rich Asymptotic Giant Branch (AGB) stars [@Srinivasan:2011; @Goldman:2017]. It is also a challenge to compare results as they are calculated based on measured values of the dust (optical constants) which can not be interpolated over. To understand the ranges and estimated errors of fitted results, they must be compared to results from different model grids.
+# Statement of need
 
-This package is designed to easily compare stellar samples and model grids for a better understanding of the results and their uncertainties. The package can be installed using `pip` and, using "entrypoints", can be accessed from any terminal prompt once installed. The fitting method uses a brute-force technique to ensure a true best fit. Grids of radiative transfer models can easily be created and added to the model grid library. The available model grids for this version are listed in
-Table 1.
+Results from these grids (e.g. luminosity, mass-loss rate) can vary dramatically as a result of the unknown properties of evolved stars, especially the oxygen-rich Asymptotic Giant Branch (AGB) stars [@Srinivasan:2011; @Goldman:2017]. It is also a challenge to compare results as they are calculated based on measured values of the dust (optical constants) which can not be interpolated over. To understand the ranges and estimated errors of fitted results, they must be compared to results from different model grids. This is particularly important given the wealth of infrared data to come from the James Webb Space Telescope (JWST).
 
-Table 1
+
+# User interface
+
+This package is designed to easily compare stellar samples and model grids for a better understanding of the results and their uncertainties. The package can be installed using `pip` and, using "entrypoints", can be accessed from any terminal prompt once installed. The fitting method uses a brute-force technique to ensure a true best fit. Grids of radiative transfer models can easily be created and added to the model grid library. The available model grids for this version are listed in Table 1.
+
+# Figures
+
+![An example of three massive oxygen-rich AGB stars in the Large Magellanic Cloud (LMC) galaxy fit with the default oxygen-rich model grid (Oss-Orich-bb). These three example sources can be fit, and this figure can be created, using the command `desk fit` and then the command `desk sed`.  ](docs/example.png)
+
+# Tables
+
+Table 1: The model grids available with this version. Shown is the name, size, chemical type of either oxygen (O) or carbon (c), the atmospheric model , and a brief description.
 
 | Grid name           | Size   | Type | Atmospheric model | Optical constants                                          | References                     |
 |---------------------|--------|------|-------------------|------------------------------------------------------------|--------------------------------|
@@ -50,8 +60,5 @@ Table 1
 | J1000-LMC           | 85,392 | C    | COMARCS           | Dust-growth grid with 1/2 solar metallicity                |                                |
 | J1000-SMC           | 85,546 | C    | COMARCS           | Dust-growth grid with 1/5 solar metallicity                |                                |
 
-# Figures
-
-![An example of three massive oxygen-rich AGB stars in the Large Magellanic Cloud (LMC) galaxy fit with the default oxygen-rich model grid (Oss-Orich-bb). These three example sources can be fit, and this figure can be created, using the command `desk fit` and then the command `desk sed`.  ](docs/example.png)
 
 # References
