+++ paper/joss_table.tex
@@ -14,7 +14,7 @@
 \vspace{3cm}
 \begin{table*}
     \centering
-    \caption{Table 1: The model grids available with this version. Shown is the name, size, chemical type of either oxygen (O) or carbon (C), the atmospheric model , and a brief description.}
+    \caption{The model grids available with this version. Shown is the name, size, chemical type of either oxygen (O) or carbon (C), the atmospheric model , and a brief description.}
     % \setlength{\tabcolsep}{0.76em}
     \begin{tabular}{ l r c c c c c}
     \hline
