### 2020-08-13 #2554 @s-goldman
> Looking at the paper I have a couple of general comments:
>
> 1. I think you should list all the available grids, with the appropriate citations, as you've done on in the online documentation. A table format might be appropriate (name, size, description + citation)?
> 2. The included figures would be a lot more useful if you included a description of what they are and how they were generated using desk. I think the latter is already included in the online documentation also.
>
> Sorry this review is a little drip fed, a necessity of time constraints unfortunately.

Thanks for these suggestions. I have added a more descriptive figure caption and a table with the grid parameters. The Markdown table format wouldn't fit, so the table is added as a png, and the latex files used to create the figure are now in the paper directory.
