### 2020-08-22 s-goldman/Dusty-Evolved-Star-Kit#147 @Deech08
[Feature Request] Requirements
As part of JOSS Review openjournals/joss-reviews#2554

While not a requirement, i think it would be useful to include the package requirements/dependencies (like astropy, seaborn, etc.) explicitly in the readme and documentation. Most useful would be on the "installation" page/section just so users will know what to expect pip to install if they don't already have it.

If you prefer not to for any reason, that's fine as well as its not a technical requirement with JOSS (since you have the package pip installable and a requirements.txt file exists).
