### 2020-08-22 s-goldman/Dusty-Evolved-Star-Kit#148 @Deech08
[BUG] Missing Dependency Install with pip
As part of JOSS Review openjournals/joss-reviews#2554

h5py not installed when running `pip install desk`

Looking at the setup.py it seems like its missing from the install requires field, along with some other packages that are listed as requirements in requirements.txt

Error is encountered when running `desk grid` following install in a new conda environment

Error message:
> (desk_rev_pip) Dhaneshs-MacBook-Pro-2:DESK_TEST dk$ desk grids
> Traceback (most recent call last):
>   File "/anaconda3/envs/desk_rev_pip/bin/desk", line 5, in <module>
>     from desk.main import main
>   File "/anaconda3/envs/desk_rev_pip/lib/python3.8/site-packages/desk/main.py", line 5, in <module>
>     from desk import console_commands
>   File "/anaconda3/envs/desk_rev_pip/lib/python3.8/site-packages/desk/console_commands.py", line 8, in <module>
>     from desk.set_up import (
>   File "/anaconda3/envs/desk_rev_pip/lib/python3.8/site-packages/desk/set_up/get_models.py", line 6, in <module>
>     import h5py
> ModuleNotFoundError: No module named 'h5py'

Conda environment details:
new environment with python 3.8, then run only `pip install desk`

Edit: Additional packages are also missing and should be added
-matplotlib, sklearn,
