### 2020-08-22 s-goldman/Dusty-Evolved-Star-Kit#150 @Deech08
[Request] Documentation: Model Grid File-size estimates
As part of JOSS Review openjournals/joss-reviews#2554

While testing out different model grids, noticing that some of these grids are quite large (currently at 1.5GB with just 3 extra model grids downloaded).

It would be best to include a page in the documentation (probably [here?](https://dusty-evolved-star-kit.readthedocs.io/en/latest/grids.html) ) that lists explicitly how big each model grid is so users can know what to expect if they plan to use many models. There should also be somewhere a warning of what the full package size, with all model grids downloaded, would be, noting that by default most are not downloaded.
