### 2020-08-22 s-goldman/Dusty-Evolved-Star-Kit#151 @Deech08
[Request] Documentation: Additional API docs / example
As part of JOSS Review openjournals/joss-reviews#2554

The command line tool works great! But if someone wanted to integrate this package into an existing code base or pipeline from within python, its not immediately obvious how to do so (i.e., what they should import, what order to run things, etc.) It would be nice to have a page in the Docs that shows how to do the same example as done in the command line version, but invoked from within python. This would include package imports, and lines of code that provide the results that are ultimately displayed and saved in the command line tool.

Edit: It would also be nice to have an example of how the plots are made (in particular, how the model can be re-computed using the fit results or something similar).
