### 2020-08-22 s-goldman/Dusty-Evolved-Star-Kit#152 @Deech08
[Request] Documentation: Additional API docs - short descriptions
As part of JOSS Review openjournals/joss-reviews#2554

Many/most modules in the subpackages have short descriptions of their functionality already, but some do not.
For example desk.fitting.dusty_fit.fit_single_source just says "short description" which I think may have just been filler text.

Another is desk.outputs.parameter_ranges.create_par

and a few in desk.set_up_package
