### 2020-08-23 s-goldman/Dusty-Evolved-Star-Kit#153 @Deech08
[Clarify] Documentation - Multiprocessing significance
As part of JOSS Review openjournals/joss-reviews#2554

From what I can gauge, multiprocessing is only used if desk fit is run on a folder with many input files. Clarification in the documentation on this would be best, letting users know that multiprocessing will only affect performance when fitting many input sets in a folder, as opposed to just make a single fit run faster.
