### 2020-09-05 #2554 @arfon
@s-goldman - looking at the cross-linked issues above, it seems like you've had a bunch more feedback. How are you getting along with your responses to these?
