### 2020-09-05 #2554 @s-goldman
@arfon Thank you for pointing this out. I wasn't getting notifications for the new issues, but I've changed that. I'll get right on these issues.
