### 2020-09-08 c84ce6b
Added documentation specifying multiprocessing has little effect on small samples
