### 2020-09-08 s-goldman/Dusty-Evolved-Star-Kit#153 @s-goldman
Documentation has been added.

commit c84ce6bea9a453a3f52c2a1e0f767860ca370e33
