### 2020-09-08 15bcb5d
Added doc strings where they were missing in set_up package.
