+++ desk/set_up/error_messages.py
@@ -2,12 +2,18 @@
 
 
 class Fitting_Range_Error(ValueError):
+    """Specific error message for bad input range"""
+
     pass
 
 
 class BadFilenameError(ValueError):
+    """Specific error message for bad input source (filename)"""
+
     pass
 
 
 class BadSourceDirectoryError(ValueError):
+    """Specific error message for bad input source (directory)"""
+
     pass
