### 2020-09-09 b05ba32
added functionality to use DESK within python environment.
