### 2020-09-11 s-goldman/Dusty-Evolved-Star-Kit#160 @Deech08
[Docs] Testing Instructions / Guide
As part of openjournals/joss-reviews#2554

Looks like you have tests, which I'm able to run. and all pass (test_console_commands.py). But I had to dig through to find these and run. Since the repository doesn't use some kind of continuous integration (Travis CI, etc.) it would be best to include a line in the readme or a small section in the documentation that tells a user how they can run the test to make sure the package installed properly and such.
