### 2020-09-11 s-goldman/Dusty-Evolved-Star-Kit#161 @Deech08
[Request] Issue Templates
As part of openjournals/joss-reviews#2554

In your docs, you have a nice page for contributing and how to do it. It would be helpful if you can make issue templates for bug reports and feature requests that list what you ask in the docs, but just lays it out within a GitHub template so its more clear and obvious for users.

(you can do this in settings, features, issues)
