### 2020-09-11 s-goldman/Dusty-Evolved-Star-Kit#160 @s-goldman
@Deech08 There is continuous integration testing done through Github actions (The link is the [python package: passing] button on the readme; [https://github.com/s-goldman/Dusty-Evolved-Star-Kit/actions](https://github.com/s-goldman/Dusty-Evolved-Star-Kit/actions)). I will make this more clear, and add some documentation about how to initiate the tests manually.
