### 2020-09-14 s-goldman/Dusty-Evolved-Star-Kit#161 @s-goldman
Completed in commit 1d1efd5517209d69cc89961183fe5de8a78123d5
