### 2020-09-14 71cbaea
Added option to specify multiprocessing core number
