### 2020-09-14 d8c92e2
removed bare except
