### 2020-09-16 450169b
Paper statement about model grid developement altered.
