### 2020-09-16 #2554 @s-goldman
> @s-goldman - a minor comment on your paper: The last sentence states its easy to create and add model grids, but it doesn't seem like this is a current feature, but rather something that you plan to implement in the future. It may be best to state that more clearly.
>
> Other than the last couple minor issues I opened, i think you've addressed most of my other concerns. I'll check through all the latest code/docs in the next few days and should be finished with my review then!

Fixed in Issue [#164](https://github.com/s-goldman/Dusty-Evolved-Star-Kit/pull/164)

All other issues opened for the JOSS review (up-to-this-point) have been addressed.
