### 2020-09-22 #2554 @TomGoffrey
@arfon Apologies, I'd taken some time off before the start of term. I think we should be able to get this completed in a day or so.

@s-goldman Which branch should I be using to generate the latest version of the paper?
