### 2020-09-22 5abef82
updated JOSS paper with table and language describing creation of model grids
