+++ paper.bib
@@ -1,3 +1,33 @@
+@Article{Aringer:2016,
+       author = {{Aringer}, B. and {Girardi}, L. and {Nowotny}, W. and {Marigo}, P. and
+         {Bressan}, A.},
+        title = "{Synthetic photometry for M and K giants and stellar evolution: hydrostatic dust-free model atmospheres and chemical abundances}",
+      journal = {\mnras},
+     keywords = {molecular data, stars: AGB and post-AGB, stars: atmospheres, stars: evolution, Hertzsprung-Russell and colour-magnitude diagrams, stars: late-type, Astrophysics - Solar and Stellar Astrophysics},
+         year = "2016",
+        month = "Apr",
+       volume = {457},
+       number = {4},
+        pages = {3611-3628},
+          doi = {10.1093/mnras/stw222},
+archivePrefix = {arXiv},
+       eprint = {1601.07025},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2016MNRAS.457.3611A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+@Article{Ossenkopf:1992,
+  author   = {{Ossenkopf}, V. and {Henning}, T. and {Mathis}, J.~S.},
+  title    = {{Constraints on cosmic silicates.}},
+  journal  = {\aap},
+  year     = {1992},
+  volume   = {261},
+  pages    = {567-578},
+  month    = Aug,
+  adsnote  = {Provided by the SAO/NASA Astrophysics Data System},
+  adsurl   = {https://ui.adsabs.harvard.edu/#abs/1992A&A...261..567O},
+  keywords = {Interstellar Matter, Silicates, Stellar Envelopes, Astronomical Models, Iron, Magnetite, Opacity, Silicon Carbides, Astrophysics},
+}
 @article{Elitzur:2001,
   author   = {{Elitzur}, Moshe and {Ivezi{\'c}}, {\v{Z}}eljko},
   title    = {{Dusty winds - I. Self-similar solutions}},
