### 2020-09-22 dc6012c
Update paper.md
