+++ paper.md
@@ -36,6 +36,7 @@ This package is designed to easily compare stellar samples and model grids for a
 # Figures
 
 ![An example of three massive oxygen-rich AGB stars in the Large Magellanic Cloud (LMC) galaxy fit with the default oxygen-rich model grid (Oss-Orich-bb). These three example sources can be fit, and this figure can be created, using the command `desk fit` and then the command `desk sed`.  ](docs/example.png)
+
 ![](paper/joss_table.png)
 
 
