### 2020-09-22 #2554 @s-goldman
@TomGoffrey The paper.md file on the main branch is now the most up-to-date.
