### 2020-09-22 #2554 @TomGoffrey
Thanks, in case it's useful in future:

`@whedon generate pdf from branch origin/<some_branch_name>`

generates the paper from a given branch.

I'll take a look tomorrow.
