### 2020-09-24 s-goldman/Dusty-Evolved-Star-Kit#165 @TomGoffrey
AssertionError if no arguments passed
**Describe the bug**
Invoking desk from the command line without any option raises an AssertionError

**To Reproduce**
```
$ desk
Traceback (most recent call last):
  File "/usr/local/bin/desk", line 8, in <module>
    sys.exit(main())
  File "/Users/tom/dev/Dusty-Evolved-Star-Kit/desk/main.py", line 36, in main
    assert False, "Invalid subparser! This should be impossible..."
AssertionError: Invalid subparser! This should be impossible...

If you suspect this is an IPython 7.12.0 bug, please report it at:
    https://github.com/ipython/ipython/issues
or send an email to the mailing list at ipython-dev@python.org

You can print a more detailed traceback right now with "%tb", or use "%debug"
to interactively debug it.

Extra-detailed tracebacks for bug-reporting purposes can be enabled via:
    %config Application.verbose_crash=True
```
**Expected behavior**

Perhaps the most sensible thing here would be to return the same as `desk --help`?

**Screenshots**

**Additional context**

 - Part of openjournals/joss-reviews#2554
 - Current master, dc6012cd.
