### 2020-09-24 s-goldman/Dusty-Evolved-Star-Kit#165 @TomGoffrey
It looks like the case of an invalid argument is handled correctly though:

```
desk abc
usage: desk [-h] {fit,grids,save_model,sed,sed_indiv,version} ...
desk: error: argument subparser_name: invalid choice: 'abc' (choose from 'fit', 'grids', 'save_model', 'sed', 'sed_indiv', 'version')
```
