### 2020-09-24 s-goldman/Dusty-Evolved-Star-Kit#168 @TomGoffrey
RuntimeError: OrderedDict mutated during iteration during desk sed
**Describe the bug**

`desk sed` produces `RuntimeError: OrderedDict mutated during iteration`. This command works fine on my OS X laptop, but fails (as shown below) on my Ubuntu desktop.

**To Reproduce**

```
$ desk fit --source='desk/put_target_data_here/MSX_LMC_807.csv' --grid='oxygen' --n=50 --distance=10

You already have the grid!

Scaling to full grid (100,000 models)

             Target: MSX_LMC_807		1/1
--------------------------------------------------------
Luminosity			|	1,000 Msun
Optical depth (at 10 um)	|	3.16
Expansion velocity (scaled)	|	5.96 km/s
Gas mass loss (scaled)		|	3.41E-06 Msun/yr
--------------------------------------------------------
tom@tom-pc:~/dev/Dusty-Evolved-Star-Kit$ desk sed
Downloading https://hpiers.obspm.fr/iers/bul/bulc/Leap_Second.dat
|===========================================================================================================================================================| 1.3k/1.3k (100.00%)         0s
Traceback (most recent call last):
  File "/home/tom/.local/bin/desk", line 11, in <module>
    sys.exit(main())
  File "/home/tom/dev/Dusty-Evolved-Star-Kit/desk/main.py", line 33, in main
    func(**funcargs)
  File "/home/tom/dev/Dusty-Evolved-Star-Kit/desk/console_commands.py", line 63, in sed
    plotting_seds.create_fig(source_path, source_filename, dest_path, save_name)
  File "/home/tom/dev/Dusty-Evolved-Star-Kit/desk/outputs/plotting_seds.py", line 197, in create_fig
    input_file = Table.read(source_path + "/" + source_filename)
  File "/home/tom/.local/lib/python3.6/site-packages/astropy/table/connect.py", line 52, in __call__
    out = registry.read(cls, *args, **kwargs)
  File "/home/tom/.local/lib/python3.6/site-packages/astropy/io/registry.py", line 520, in read
    'read', cls, path, fileobj, args, kwargs)
  File "/home/tom/.local/lib/python3.6/site-packages/astropy/io/registry.py", line 595, in _get_valid_format
    valid_formats = identify_format(mode, cls, path, fileobj, args, kwargs)
  File "/home/tom/.local/lib/python3.6/site-packages/astropy/io/registry.py", line 414, in identify_format
    for data_format, data_class in _identifiers:
RuntimeError: OrderedDict mutated during iteration

If you suspect this is an IPython 7.16.1 bug, please report it at:
    https://github.com/ipython/ipython/issues
or send an email to the mailing list at ipython-dev@python.org

You can print a more detailed traceback right now with "%tb", or use "%debug"
to interactively debug it.

Extra-detailed tracebacks for bug-reporting purposes can be enabled via:
    %config Application.verbose_crash=True
```

**Expected behavior**
Command produces plot saved in file.

**Screenshots**
If applicable, add screenshots to help explain your problem.

**Additional context**
 - Part of openjournals/joss-reviews#2554
 - Current master, dc6012c.
 - Python 3.6.9, Ubuntu 18.04.05.
 - Connected to desktop over ssh. I doubt it's relevant, but I'll mention it anyway.
