### 2020-09-24 s-goldman/Dusty-Evolved-Star-Kit#168 @TomGoffrey
FWIW the laptop that this does run on is running python 3.7.6, so I'm wondering if there's a 3.6 vs 3.7 change in behaviour?
