### 2020-09-24 #2554 @TomGoffrey
I've raised a few issues I came across whilst looking through this.

Other than that my only other comment is to come back to the discussion of citing previous use. In your readme you state:

> The method used is similar to that of [Goldman et al. 2017]...

The docs for JOSS ("What should my paper contain?") state:

> Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it.

My understanding is that Goldman et al. 2017 used an early version of the code, but that's not stated in the paper, so you may wish to include a comment to that effect.
