### 2020-09-24 #2554 @s-goldman
@TomGoffrey Thanks for the issues and for really testing the limits of the code. I'll try to have these done soon.

I was a little unsure on how to phrase how the code was previously used, and was hoping you might give some guidance on it, so thank you. I will add that an earlier version of the code has been used in Orosz et al. 2017 & Goldman et al. 2017, 2018, 2019a, 2019b.
