### 2020-09-24 #2554 @TomGoffrey
No worries. I think something along the lines of

> Early versions of this package have been used in X, Y, Z

should be fine.

I'm still trying to work out if the OrderedDict error is specific to my set-up, or a python3.6 vs python3.7 issue. I know there were changes to Dict vs OrderedDict between those two versions, but I'm not aware of those changes causing such an issue.
