### 2020-09-24 #2554 @s-goldman
> I'm still trying to work out if the OrderedDict error is specific to my set-up, or a python3.6 vs python3.7 issue. I know there were changes to Dict vs OrderedDict between those two versions, but I'm not aware of those changes causing such an issue.

@TomGoffrey I appreciate that, i'll try to see if I can reproduce the error on my end.
