### 2020-09-24 d15f1ec
Update paper.md
