+++ paper.bib
@@ -28,6 +28,18 @@ archivePrefix = {arXiv},
   adsurl   = {https://ui.adsabs.harvard.edu/#abs/1992A&A...261..567O},
   keywords = {Interstellar Matter, Silicates, Stellar Envelopes, Astronomical Models, Iron, Magnetite, Opacity, Silicon Carbides, Astrophysics},
 }
+@Article{Orosz2017,
+  author   = {{Orosz}, G. and {Imai}, H. and {Dodson}, R. and {Rioja}, M.~J. and {Frey}, S. and {Burns}, R.~A. and {Etoka}, S. and {Nakagawa}, A. and {Nakanishi}, H. and {Asaki}, Y. and {Goldman}, S.~R. and {Tafoya}, D.},
+  title    = {{Astrometry of OH/IR Stars Using 1612 MHz Hydroxyl Masers. I. Annual Parallaxes of WX Psc and OH138.0+7.2}},
+  journal  = {\aj},
+  year     = {2017},
+  volume   = {153},
+  month    = Mar,
+  adsnote  = {Provided by the SAO/NASA Astrophysics Data System},
+  adsurl   = {https://ui.adsabs.harvard.edu/#abs/2017AJ....153..119O},
+  doi      = {10.3847/1538-3881/153/3/119},
+  keywords = {astrometry, masers, stars: AGB and post-AGB, stars: individual: WX Psc, OH 138.0+7.2, techniques: interferometric, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Astrophysics of Galaxies, Astrophysics - Instrumentation and Methods for Astrophysics},
+}
 @article{Elitzur:2001,
   author   = {{Elitzur}, Moshe and {Ivezi{\'c}}, {\v{Z}}eljko},
   title    = {{Dusty winds - I. Self-similar solutions}},
@@ -113,9 +125,52 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2017MNRAS.465..403G},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-
-
-
+@Article{Goldman2018,
+  author   = {{Goldman}, Steven R. and {van Loon}, Jacco Th. and {G{\'o}mez}, Jos{\'e} F. and {Green}, James A. and {Zijlstra}, Albert A. and {Nanni}, Ambra and {Imai}, Hiroshi and {Whitelock}, Patricia A. and {Groenewegen}, Martin A.~T. and {Oliveira}, Joana M.},
+  title    = {{A dearth of OH/IR stars in the Small Magellanic Cloud}},
+  journal  = {\mnras},
+  year     = {2018},
+  volume   = {473},
+  pages    = {3835-3853},
+  month    = Jan,
+  adsnote  = {Provided by the SAO/NASA Astrophysics Data System},
+  adsurl   = {https://ui.adsabs.harvard.edu/#abs/2018MNRAS.473.3835G},
+  doi      = {10.1093/mnras/stx2601},
+  keywords = {masers, stars: AGB and post-AGB, stars: mass-loss, supergiants, stars: winds, outflows, Magellanic Clouds, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Astrophysics of Galaxies},
+}
+@Article{Goldman2019a,
+  author        = {Goldman, S. R. and Boyer, M. L. and McQuinn, K. B. W. and Whitelock, P. A. and McDonald, I. and van Loon, J. T. and Skillman, E. D. and Gehrz, R. D. and Javadi, A. and Sloan, G. C. and Jones, O. C. and Groenewegen, M. A. T. and Menzies, J. W.},
[…]
