+++ paper.bib
@@ -28,7 +28,7 @@ archivePrefix = {arXiv},
   adsurl   = {https://ui.adsabs.harvard.edu/#abs/1992A&A...261..567O},
   keywords = {Interstellar Matter, Silicates, Stellar Envelopes, Astronomical Models, Iron, Magnetite, Opacity, Silicon Carbides, Astrophysics},
 }
-@Article{Orosz2017,
+@Article{Orosz:2017,
   author   = {{Orosz}, G. and {Imai}, H. and {Dodson}, R. and {Rioja}, M.~J. and {Frey}, S. and {Burns}, R.~A. and {Etoka}, S. and {Nakagawa}, A. and {Nakanishi}, H. and {Asaki}, Y. and {Goldman}, S.~R. and {Tafoya}, D.},
   title    = {{Astrometry of OH/IR Stars Using 1612 MHz Hydroxyl Masers. I. Annual Parallaxes of WX Psc and OH138.0+7.2}},
   journal  = {\aj},
@@ -125,7 +125,7 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2017MNRAS.465..403G},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-@Article{Goldman2018,
+@Article{Goldman:2018,
   author   = {{Goldman}, Steven R. and {van Loon}, Jacco Th. and {G{\'o}mez}, Jos{\'e} F. and {Green}, James A. and {Zijlstra}, Albert A. and {Nanni}, Ambra and {Imai}, Hiroshi and {Whitelock}, Patricia A. and {Groenewegen}, Martin A.~T. and {Oliveira}, Joana M.},
   title    = {{A dearth of OH/IR stars in the Small Magellanic Cloud}},
   journal  = {\mnras},
@@ -138,7 +138,7 @@ archivePrefix = {arXiv},
   doi      = {10.1093/mnras/stx2601},
   keywords = {masers, stars: AGB and post-AGB, stars: mass-loss, supergiants, stars: winds, outflows, Magellanic Clouds, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Astrophysics of Galaxies},
 }
-@Article{Goldman2019a,
+@Article{Goldman:2019a,
   author        = {Goldman, S. R. and Boyer, M. L. and McQuinn, K. B. W. and Whitelock, P. A. and McDonald, I. and van Loon, J. T. and Skillman, E. D. and Gehrz, R. D. and Javadi, A. and Sloan, G. C. and Jones, O. C. and Groenewegen, M. A. T. and Menzies, J. W.},
   title         = {aaaa An Infrared Census of DUST in Nearby Galaxies with Spitzer (DUSTiNGS). V. The Period{\ndash}Luminosity Relation for Dusty Metal-poor AGB Stars},
   journal       = {ApJ},
@@ -155,7 +155,7 @@ archivePrefix = {arXiv},
   primaryclass  = {astro-ph.SR},
   url           = {https://ui.adsabs.harvard.edu/abs/2019ApJ...877...49G},
 }
-@Article{Goldman2019b,
+@Article{Goldman:2019b,
   author        = {Goldman, S. R. and Boyer, M. L. and McQuinn, K. B. and Sloan, G. C. and McDonald, I. and van Loon, J. T. and Zijlstra, A. A. and Hirschauer, A. S. and Skillman, E. D. and Srinivasan, S.},
   title         = {AGB Stars in the Nearby Dwarf Galaxy: Leo P},
   journal       = {ApJ},
