### 2020-09-28 da44a18
Upon entering only desk'', the command line will return the desk --help'' response.
