### 2020-09-28 e935e86
Using function and not command line.
