### 2020-09-28 f7efd26
Using function and not command line.
