### 2020-09-28 6489eb9
Error message for multiprocessing = -1, and 0 cores
