### 2020-09-28 6b984d7
Error message for multiprocessing = -1, and 0 cores
