### 2020-09-28 3a96df2
Error message for multiprocessing = -1, and 0 cores
