### 2020-09-28 s-goldman/Dusty-Evolved-Star-Kit#168 @s-goldman
@TomGoffrey Hmmm, what version of astropy are you using?
