### 2020-09-29 s-goldman/Dusty-Evolved-Star-Kit#168 @TomGoffrey
astropy version is `'4.0.1.post1'`

The error repeats every time I try from command line.

Some slight weirdness from trying to trigger the issue in ipython3 (by running `import desk;desk.sed()`), the first time fails in the same manner as before, but the second time I run it appears to work fine. This pattern seems repeatable too.

I'm leaning towards something being not quite right with the machine I'm testing on. I'll have physical access to it tomorrow, so I'll try and investigate more then. If anything I can at least restart it for the first time in 250 days, lockdown hasn't been kind to it.
