### 2020-09-29 s-goldman/Dusty-Evolved-Star-Kit#168 @TomGoffrey
In case it's useful, here's the output from running the commands in ipython3. There's some complaints at the end about running `sys.exit()` but the .png file is clearly new.

```python
In [1]: import desk

In [2]: desk.sed()
RuntimeError                              Traceback (most recent call last)
<ipython-input-2-7fde54a9a56d> in <module>
----> 1 desk.sed()

~/dev/Dusty-Evolved-Star-Kit/desk/console_commands.py in sed(source_path, source_filename, dest_path, save_name)
     61         SED figure with data in blue and model in black.
     62     """
---> 63     plotting_seds.create_fig(source_path, source_filename, dest_path, save_name)
     64
     65

~/dev/Dusty-Evolved-Star-Kit/desk/outputs/plotting_seds.py in create_fig(source_path, source_filename, dest_path, save_name)
    195     """
    196
--> 197     input_file = Table.read(source_path + "/" + source_filename)
    198     n = len(input_file)  # number of fit sources
    199

~/.local/lib/python3.6/site-packages/astropy/table/connect.py in __call__(self, *args, **kwargs)
     50     def __call__(self, *args, **kwargs):
     51         cls = self._cls
---> 52         out = registry.read(cls, *args, **kwargs)
     53
     54         # For some readers (e.g., ascii.ecsv), the returned `out` class is not

~/.local/lib/python3.6/site-packages/astropy/io/registry.py in read(cls, format, *args, **kwargs)
    518
    519             format = _get_valid_format(
--> 520                 'read', cls, path, fileobj, args, kwargs)
    521
    522         reader = get_reader(format, cls)

~/.local/lib/python3.6/site-packages/astropy/io/registry.py in _get_valid_format(mode, cls, path, fileobj, args, kwargs)
    593     """
    594
--> 595     valid_formats = identify_format(mode, cls, path, fileobj, args, kwargs)
    596
    597     if len(valid_formats) == 0:

~/.local/lib/python3.6/site-packages/astropy/io/registry.py in identify_format(origin, data_class_required, path, fileobj, args, kwargs)
    412     """
    413     valid_formats = []
--> 414     for data_format, data_class in _identifiers:
    415         if _is_best_match(data_class_required, data_class, _identifiers):
    416             if _identifiers[(data_format, data_class)](

RuntimeError: OrderedDict mutated during iteration

In [3]: desk.sed()
An exception has occurred, use %tb to see the full traceback.

SystemExit

/home/tom/.local/lib/python3.6/site-packages/IPython/core/interactiveshell.py:3351: UserWarning: To exit: use 'exit', 'quit', or Ctrl-D.
  warn("To exit: use 'exit', 'quit', or Ctrl-D.", stacklevel=1)

In [4]: ls -l *.png
-rw------- 1 tom tom 56750 Sep 29 10:40 output_sed_MSX_LMC_807.png

In [5]: from datetime import datetime;datetime.now().strftime("%d/%m/%Y %H:%M:%S")
Out[5]: '29/09/2020 10:40:48'
```
