### 2020-09-29 s-goldman/Dusty-Evolved-Star-Kit#168 @s-goldman
Well, I've figured out the second system exit issue, but not the initial mutated Ordered dictionary issue. The code was working, but the output text for the python interface was more jarring than was intended. Thanks for finding this and pointing it out. I've fixed it in #172.
