### 2020-09-29 s-goldman/Dusty-Evolved-Star-Kit#168 @TomGoffrey
Thanks, the second call to `desk.sed()` runs cleanly for me now. I'll update tomorrow if/when I make progress on the original issue.
