### 2020-09-30 e7b8693
More flexible method of creating output file.
