### 2020-09-30 c55f25c
Better error message when fitting_results.csv is missing.
