### 2020-09-30 s-goldman/Dusty-Evolved-Star-Kit#168 @TomGoffrey
OK, after a solid 6 hours of debugging and arguing with Docker, I've tracked down the issue on the problematic machine to a rogue installation of an old version of asdf living in `/usr/lib/python3/dist-packages`.

Sorry for the false alarm.
