### 2020-09-30 #2554 @TomGoffrey
OK, so my last remaining point is a question over the references in the table. I think these ought to be in the references for the paper, but they're missing.

I'm guessing this is related to the fact you've had to include an image of the table rather than include it directly as markdown. Could I clarify what the issue with the markdown for the table was? Did it just not fit on the page?
