### 2020-09-30 s-goldman/Dusty-Evolved-Star-Kit#168 @s-goldman
@TomGoffrey Thanks for putting in the work to track it down!
