### 2020-09-30 0876080
Updated paper text and bib including refs
