+++ paper.bib
@@ -16,6 +16,59 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2016MNRAS.457.3611A},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+@ARTICLE{Begemann:1997,
+       author = {{Begemann}, B. and {Dorschner}, J. and {Henning}, Th. and
+         {Mutschke}, H. and {G{\"u}rtler}, J. and {K{\"o}mpe}, C. and {Nass}, R.},
+        title = "{Aluminum Oxide and the Opacity of Oxygen-rich Circumstellar Dust in the 12-17 Micron Range}",
+      journal = {\apj},
+     keywords = {Stars: Circumstellar Matter, ISM: Dust, Extinction, Infrared: ISM: Lines and Bands, Stars: Late-Type},
+         year = 1997,
+        month = feb,
+       volume = {476},
+       number = {1},
+        pages = {199-208},
+          doi = {10.1086/303597},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1997ApJ...476..199B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+@article{Elitzur:2001,
+  author   = {{Elitzur}, Moshe and {Ivezi{\'c}}, {\v{Z}}eljko},
+  title    = {{Dusty winds - I. Self-similar solutions}},
+  journal  = {\mnras},
+  year     = {2001},
+  volume   = {327},
+  pages    = {403-421},
+  month    = Oct,
+  adsnote  = {Provided by the SAO/NASA Astrophysics Data System},
+  adsurl   = {https://ui.adsabs.harvard.edu/#abs/2001MNRAS.327..403E},
+  doi      = {10.1046/j.1365-8711.2001.04706.x},
+  keywords = {STARS: AGB AND POST-AGB, CIRCUMSTELLAR MATTER, STARS: LATE-TYPE, STARS: WINDS, OUTFLOWS, DUST, EXTINCTION, INFRARED: STARS, Astrophysics},
+}
+@ARTICLE{Henning:1995,
+       author = {{Henning}, T. and {Begemann}, B. and {Mutschke}, H. and {Dorschner}, J.},
+        title = "{Optical properties of oxide dust grains.}",
+      journal = {\aaps},
+     keywords = {DUST, EXTINCTION, INFRARED: STARS, ISM: LINES AND BANDS, CIRCUMSTELLAR MATTER, STARS: AGB: POST-AGB, METHODS: LABORATORY},
+         year = 1995,
+        month = jul,
+       volume = {112},
+        pages = {143},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1995A&AS..112..143H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+@ARTICLE{Jaeger:1998,
+       author = {{Jaeger}, C. and {Molster}, F.~J. and {Dorschner}, J. and
+         {Henning}, Th. and {Mutschke}, H. and {Waters}, L.~B.~F.~M.},
+        title = "{Steps toward interstellar silicate mineralogy. IV. The crystalline revolution}",
+      journal = {\aap},
+     keywords = {LINE: IDENTIFICATION, METHODS: LABORATORY, CIRCUMSTELLAR MATTER, INFRARED: ISM: LINES AND BANDS, INFRARED: STARS},
+         year = 1998,
+        month = nov,
+       volume = {339},
+        pages = {904-916},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1998A&A...339..904J},
[…]
