+++ paper/joss_table.tex
@@ -1,7 +1,7 @@
 %% using aastex version 6.3
 \documentclass[onecolumn]{aastex63}
 % \usepackage[papersize={10in,7in}, left=0.1in,right=0.1in,top=1in,bottom=0.1in]{geometry}
-\usepackage[papersize={8.75in,5in}, top=3cm, bottom=0cm, left=2.0cm, right=0.75cm]{geometry}
+\usepackage[papersize={8.25in,4.1in}, top=2.1cm, bottom=-0.5cm, left=1.5cm, right=0.25cm]{geometry}
 
 \usepackage{graphicx}
 \usepackage{ascii}
@@ -14,33 +14,32 @@
 \vspace{3cm}
 \begin{table*}
     \centering
-    \caption{The model grids available with this version. Shown is the name, size, chemical type of either oxygen (O) or carbon (C), the atmospheric model , and a brief description.}
+    \caption{The model grids available with this version. Shown is the name, size, chemical type of either oxygen (O) or carbon (C), the atmospheric model, and a brief description.}
     % \setlength{\tabcolsep}{0.76em}
     \begin{tabular}{ l r c c c c c}
     \hline
     Grid name & Size & Type & Atmospheric model & Optical constants & References \\
     \hline \hline
-    Oss-Orich-aringer & 2,000 & O & COMARCS & Warm silicates & 1, 8 \\
-    Oss-Orich-bb & 2,000 & O & Black body (BB) & Warm silicates & 8\\
-    Crystalline-20-bb & 2,000 & O & BB & 80\% warm silicates, 20\% crystalline silicates & 6, 8  \\
-    corundum-20-bb & 2,000 & O & BB & 80\% warm silicates, 20\% corundum silicates &  2, 8\\
-    big-grain & 2,000 & O & BB & Warm silicates with higher maximum dust grain size of 0.35 &  8\\
-    fifth-iron & 500 & O & BB & 80\% warm silicates, 20\% iron grains &  5, 8\\
-    half-iron & 500 & O & BB & 50\% warm silicates, 50\% iron grains &  5, 8\\
-    one-fifth-carbon & 500 & O & BB & 80\% warm silicates, 20\% carbonaceous grains &  8, 9\\
-    arnold-palmer & 500 & O & BB & 50\% warm silicates, 50\% carbonaceous grains &  8, 9\\
-    desk-mix & xx,xxx & O & MARCS & Mixture of olivine(48--88\%), corundum(48--8\%), and iron (4\%) & 2, 3, 4, 5 \\
-    Zubko-Crich-aringer & 2,000 & C & COMARCS & Amorphous carbon grains &  1, 9\\
-    Zubko-Crich-bb & 2,000 & C & BB & Amorphous carbon grains &  9\\
-    H11-LMC & 90,899 & C & COMARCS & Dust-growth grid with 1/2 solar metallicity &  7\\
-    H11-SMC & 91,058 & C & COMARCS & Dust-growth grid with 1/5 solar metallicity &  7\\
-    J1000-LMC & 85,392 & C & COMARCS & Dust-growth grid with 1/2 solar metallicity &  7\\
-    J1000-SMC & 85,546 & C & COMARCS & Dust-growth grid with 1/5 solar metallicity & 7\\
+    Oss-Orich-aringer & 2,000 & O & COMARCS & Warm silicates & 1, 6\\
+    Oss-Orich-bb & 2,000 & O & Black body (BB) & Warm silicates & 6\\
+    Crystalline-20-bb & 2,000 & O & BB & 80\% warm silicates, 20\% crystalline silicates & 4, 6\\
[…]
