### 2020-09-30 70c0a02
Updated table figure.
