+++ paper/joss_table.tex
@@ -1,7 +1,7 @@
 %% using aastex version 6.3
 \documentclass[onecolumn]{aastex63}
 % \usepackage[papersize={10in,7in}, left=0.1in,right=0.1in,top=1in,bottom=0.1in]{geometry}
-\usepackage[papersize={8.25in,4.1in}, top=2.1cm, bottom=-0.5cm, left=1.5cm, right=0.25cm]{geometry}
+\usepackage[papersize={7.85in,5.0in}, top=2.5cm, bottom=-0.5cm, left=1.0cm, right=0.25cm]{geometry}
 
 \usepackage{graphicx}
 \usepackage{ascii}
