### 2020-10-01 3a1db8a
Updated table figure.
