### 2020-10-01 a4f820e
Small paper edits.
