### 2020-10-01 50a202a
Small paper edits.
