+++ paper.bib
@@ -226,6 +226,22 @@ archivePrefix = {arXiv},
   primaryclass  = {astro-ph.SR},
   url           = {https://ui.adsabs.harvard.edu/abs/2019MNRAS.487..502N},
 }
+@ARTICLE{Wiegert:2020,
+       author = {{Wiegert}, J. and {Groenewegen}, M.~A.~T. and {Jorissen}, A. and
+         {Decin}, L. and {Danilovich}, T.},
+        title = "{How to disentangle geometry and mass-loss rate from AGB-star spectral energy distributions -- The case of EP Aqr}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Astrophysics of Galaxies},
+         year = 2020,
+        month = aug,
+          eid = {arXiv:2008.11525},
+        pages = {arXiv:2008.11525},
+archivePrefix = {arXiv},
+       eprint = {2008.11525},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv200811525W},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
 @ARTICLE{Zubko:1996,
        author = {{Zubko}, V.~G. and {Mennella}, V. and {Colangeli}, L. and
          {Bussoletti}, E.},
