### 2020-10-01 f1c70ec
Small paper edits.
