### 2020-10-01 #2554 @s-goldman
@TomGoffrey Yes, I tried to use a markdown table but it didn't really fit and I wasn't quite satisfied. Here is an
[example](https://github.com/openjournals/joss-reviews/files/5312127/po9acici7e4jtteyrufo.pdf) of what that looked like.

I am very open to reformatting the table however you Dhanesh, and Arfon think works the best. Regardless, I have added the references that I missed, reformatted the table to make it more readable, and made some small edits to the paper to improve the flow in PR [176](https://github.com/s-goldman/Dusty-Evolved-Star-Kit/pull/176)
