### 2020-10-01 #2554 @TomGoffrey
That new table looks good to me.

@arfon My review is complete, happy to recommend publication.
