### 2020-10-02 c76fcf0
Moved paper directory to docs.
