### 2020-10-02 88410c4
Moved paper directory to docs.
